--- src/App/InformasjonOmBedriftOgAnsatte/MineAnsatte/BlaIAnsatte/BlaIAnsatte.less
@@ -1,6 +1,0 @@
-@import '~nav-frontend-core/less/_variabler';
-
-
-
-
-
